% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\docType{methods}
\name{initializeAnalysis}
\alias{initializeAnalysis}
\title{constructor for}
\usage{
\S4method{initialize}{Analysis}(.Object, ..., id, name,
  description = character(0), pro)
}
\description{
This is the constructor.
}
