% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\docType{methods}
\name{initializeProcess}
\alias{initializeProcess}
\title{constructor for}
\usage{
\S4method{initialize}{Process}(.Object, ..., process, id, name,
  description = character(0), characteristics = list(0),
  data = data.frame(0))
}
\description{
This is the constructor.
}
