% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProData.R
\name{initializeProData}
\alias{initializeProData}
\title{constructor for}
\usage{
\S4method{initialize}{ProData}(
  .Object,
  ...,
  id,
  name,
  description = NA_character_,
  pro,
  characteristic,
  data
)
}
\description{
This is the constructor.
}
