% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Product.R
\name{initializeProduct}
\alias{initializeProduct}
\title{constructor for}
\usage{
\S4method{initialize}{Product}(
  .Object,
  ...,
  product,
  id,
  name,
  description = character(0),
  characteristics = list(0)
)
}
\description{
This is the constructor.
}
