% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{initializeAnalysis}
\alias{initializeAnalysis}
\title{constructor for}
\usage{
\S4method{initialize}{Analysis}(.Object, ..., id, name, description = NA_character_, data, pro)
}
\description{
This is the constructor.
}
