% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Characteristic.R
\name{initializeCharacteristic}
\alias{initializeCharacteristic}
\title{constructor for}
\usage{
\S4method{initialize}{Characteristic}(
  .Object,
  ...,
  id,
  name,
  description = character(0),
  kvalue = character(0)
)
}
\description{
This is the constructor.
}
