% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pro.R
\name{initializePro}
\alias{initializePro}
\title{constructor for}
\usage{
\S4method{initialize}{Pro}(
  .Object,
  ...,
  id,
  name,
  description = character(0),
  characteristics = list(0),
  datasets = list(0)
)
}
\description{
This is the constructor.
}
